# docs/swagger.yaml
basePath: /api
definitions:
  dto.CreateKnowledgeBaseInput:
    properties:
      dataset_id:
        type: string
      knowledgebase_name:
        type: string
    type: object
  dto.CreateUserInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
...
  dto.KnowledgeBaseListResult:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      result:
        items:
          $ref: '#/definitions/models.KnowledgeBase'
        type: array
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateKnowledgeBaseInput:
    properties:
      dataset_id:
        type: string
      knowledgebase_name:
        type: string
    type: object
  dto.UpdateUserInput:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  models.Document:
    properties:
      created_at:
        type: string
      description:
        type: string
      doc_num:
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      knowledge_base:
        $ref: '#/definitions/models.KnowledgeBase'
      knowledge_base_id:
        type: string
      name:
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.JSONMap:
    additionalProperties: true
    type: object
  models.KnowledgeBase:
    properties:
      created_at:
        type: string
      dataset_id:
        type: string
      description:
        type: string
      doc_num:
        type: integer
      document_count:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      tags:
        $ref: '#/definitions/models.JSONMap'
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      documents:
        description: 关联
        items:
          $ref: '#/definitions/models.Document'
        type: array
      email:
        type: string
      id:
        type: string
      nickname:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 用于用户增删改查的 RESTful API
  title: 用户管理系统 API
  version: "1.0"
paths:
...
  /knowledge_bases:
    get:
      consumes:
      - application/json
      description: 分页获取知识库列表
      parameters:
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 知识库列表
          schema:
            $ref: '#/definitions/dto.KnowledgeBaseListResult'
        "400":
          description: 参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取所有知识库
      tags:
      - 知识库模块
    post:
      consumes:
      - application/json
      description: 创建一个新的知识库
      parameters:
      - description: 知识库创建输入
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateKnowledgeBaseInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的知识库信息
          schema:
            $ref: '#/definitions/models.KnowledgeBase'
        "400":
          description: 请求参数错误
          schema:
            properties:
              ' details':
                type: string
              error:
                type: string
            type: object
        "500":
          description: 系统内部错误
          schema:
            properties:
              ' detail':
                type: string
              error:
                type: string
            type: object
      summary: 创建知识库
      tags:
      - 知识库模块
  /knowledge_bases/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的知识库
      parameters:
      - description: 知识库ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 知识库不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 删除知识库
      tags:
      - 知识库模块
    get:
      consumes:
      - application/json
      description: 根据ID获取知识库的详细信息
      parameters:
      - description: 知识库ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 知识库详情
          schema:
            $ref: '#/definitions/models.KnowledgeBase'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 知识库不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取知识库详情
      tags:
      - 知识库模块
    put:
      consumes:
      - application/json
      description: 更新知识库信息
      parameters:
      - description: 知识库ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      - description: 更新输入
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateKnowledgeBaseInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的知识库
          schema:
            $ref: '#/definitions/models.KnowledgeBase'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 知识库不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 更新知识库
      tags:
      - 知识库模块
  /users:
    get:
      consumes:
      - application/json
      description: 获取系统中所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户列表
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/models.User'
                type: array
            type: object
      security:
      - Bearer: []
      summary: 获取所有用户
      tags:
      - 用户模块
    post:
      consumes:
      - application/json
      description: 接收用户信息并创建用户记录
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回用户信息
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Bearer: []
      summary: 创建新用户
      tags:
      - 用户模块
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - 用户模块
    get:
      consumes:
      - application/json
      description: 根据用户ID获取特定用户的详细信息
      parameters:
      - description: 用户ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户详细信息
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 用户不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 获取单个用户信息
      tags:
      - 用户模块
    put:
      consumes:
      - application/json
      description: 更新用户的各种信息，支持部分字段更新
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户更新信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Bearer: []
      summary: 更新用户信息
      tags:
      - 用户模块
securityDefinitions:
  Bearer:
    description: '请输入 ''Bearer '' + JWT 令牌，例如: "Bearer abcde12345"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"